+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-### 0.9.1
+## 0.9.2
+ - Added the option to drag nodes as a child of an empty node, by hovering over the far right of it.
+ - Added child option to placement enum
+
+## 0.9.1
  - Updated Documentation.
  - Made internal changes for performance.